banish update and inverse transform
