+++ aeon/transformations/collection/compose/_identity.py
@@ -1,16 +1,16 @@
 """Identity transformer."""
 
+from aeon.transformations.base import InverseTransformerMixin
 from aeon.transformations.collection import BaseCollectionTransformer
 from aeon.utils.data_types import COLLECTIONS_DATA_TYPES
 
 
-class CollectionId(BaseCollectionTransformer):
+class CollectionId(BaseCollectionTransformer, InverseTransformerMixin):
     """Identity transformer, returns data unchanged in transform/inverse_transform."""
 
     _tags = {
         "X_inner_type": COLLECTIONS_DATA_TYPES,
         "fit_is_empty": True,
-        "capability:inverse_transform": True,
         "capability:multivariate": True,
         "capability:unequal_length": True,
         "capability:missing_values": True,
