+++ aeon/transformations/series/base.py
@@ -111,9 +111,8 @@ class BaseSeriesTransformer(BaseSeriesEstimator, BaseTransformer):
         if y is not None:
             self._check_y(y)
 
-        # #2768
-        # if not fit_empty:
-        #     self._check_shape(X)
+        if not fit_empty:
+            self._check_shape(X)
 
         Xt = self._transform(X, y)
         return self._postprocess_series(Xt, axis=axis)
@@ -165,86 +164,22 @@ class BaseSeriesTransformer(BaseSeriesEstimator, BaseTransformer):
         self.is_fitted = True
         return self._postprocess_series(Xt, axis=axis)
 
-    @final
-    def inverse_transform(self, X, y=None, axis=1):
-        """Inverse transform X and return an inverse transformed version.
-
-        State required:
-             Requires state to be "fitted".
-
-        Parameters
-        ----------
-        X : Input data
-            Data to fit transform to, of valid collection type.
-        y : Target variable, default=None
-             Additional data, e.g., labels for transformation
-        axis : int, default = 1
-            Axis of time in the input series.
-            If ``axis == 0``, it is assumed each column is a time series and each row is
-            a time point. i.e. the shape of the data is ``(n_timepoints,
-            n_channels)``.
-            ``axis == 1`` indicates the time series are in rows, i.e. the shape of
-            the data is ``(n_channels, n_timepoints)`.``axis is None`` indicates
-            that the axis of X is the same as ``self.axis``.
-
-        Returns
-        -------
-        inverse transformed version of X
-            of the same type as X
-        """
-        if not self.get_tag("capability:inverse_transform"):
-            raise NotImplementedError(
-                f"{type(self)} does not implement inverse_transform"
-            )
-
-        # check whether is fitted
-        self._check_is_fitted()
-        X = self._preprocess_series(X, axis=axis, store_metadata=False)
-        Xt = self._inverse_transform(X=X, y=y)
-        return self._postprocess_series(Xt, axis=axis)
-
-    @final
-    def update(self, X, y=None, update_params=True, axis=1):
-        """Update transformer with X, optionally y.
-
-        Parameters
-        ----------
-        X : data to update of valid series type.
-        y : Target variable, default=None
-            Additional data, e.g., labels for transformation
-        update_params : bool, default=True
-            whether the model is updated. Yes if true, if false, simply skips call.
-            argument exists for compatibility with forecasting module.
-        axis : int, default=None
-            axis along which to update. If None, uses self.axis.
-
-        Returns
-        -------
-        self : a fitted instance of the estimator
-        """
-        # check whether is fitted
-        self._check_is_fitted()
-        X = self._preprocess_series(X, axis, False)
-        return self._update(X=X, y=y, update_params=update_params)
-
     def _fit(self, X, y=None):
         """Fit transformer to X and y.
 
         private _fit containing the core logic, called from fit
 
+        A default implementation for the fit_is_empty tag. Other transformers
+        should override this.
+
         Parameters
         ----------
         X : Input data
             Data to fit transform to, of valid collection type.
         y : Target variable, default=None
             Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        self: a fitted instance of the estimator
         """
-        # default fit is "no fitting happens"
-        return self
+        ...
 
     @abstractmethod
     def _transform(self, X, y=None):
@@ -263,6 +198,7 @@ class BaseSeriesTransformer(BaseSeriesEstimator, BaseTransformer):
         -------
         transformed version of X
         """
+        ...
 
     def _fit_transform(self, X, y=None):
         """Fit to data, then transform it.
@@ -270,6 +206,9 @@ class BaseSeriesTransformer(BaseSeriesEstimator, BaseTransformer):
         Fits the transformer to X and y and returns a transformed version of X.
 
         private _fit_transform containing the core logic, called from fit_transform.
+
+        Non-optimised default implementation; override when a better
+        method is possible for a given algorithm.
 
         Parameters
         ----------
@@ -282,35 +221,9 @@ class BaseSeriesTransformer(BaseSeriesEstimator, BaseTransformer):
         -------
         transformed version of X.
         """
-        # Non-optimized default implementation; override when a better
-        # method is possible for a given algorithm.
-        self._fit(X, y)
+        if not self.get_tag("fit_is_empty"):
+            self._fit(X, y)
         return self._transform(X, y)
-
-    def _inverse_transform(self, X, y=None):
-        """Inverse transform X and return an inverse transformed version.
-
-        private _inverse_transform containing core logic, called from inverse_transform.
-
-        Parameters
-        ----------
-        X : Input data
-            Time series to fit transform to, of valid collection type.
-        y : Target variable, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        inverse transformed version of X
-            of the same type as X.
-        """
-        raise NotImplementedError(
-            f"{self.__class__.__name__} does not support inverse_transform"
-        )
-
-    def _update(self, X, y=None, update_params=True):
-        # standard behaviour: no update takes place, new data is ignored
-        return self
 
     def _postprocess_series(self, Xt, axis):
         """Postprocess data Xt to revert to original shape.
